New format; move up; contract + todo * Merge GetAssociateClan and UpdateAssociateClan into AssociateClan * Flip UpdateTag * Move ProcessStep to upper level; may add more factories like ProcessStep.doNothing (perhaps setClan(String), etc.) * Fix map restrictions (allow args map to accept any type; this is still an implementation detail but also should make adding additional, off-spec fields/values at compile possible * TODO: Clan ID representation; cancel Bio move mentioned in last commit
